Project/shodan_tool/shodan_nmap/requirements.txt
shodan==1.27.0
python-nmap==0.6.1
python-dotenv==1.0.0